亚像素角点检测以及GUI显示

+++ cornerdialog.h
@@ -16,7 +16,7 @@ class CornerDialog : public QDialog
     Q_OBJECT
     
 public:
-    explicit CornerDialog(Mat &img,vector<Point2d> &corners, QWidget *parent = 0);
+    explicit CornerDialog(Mat &img,vector<Point2f> &corners, QWidget *parent = 0);
     ~CornerDialog();
     void wheelEvent(QWheelEvent *);
     void mouseDoubleClickEvent(QMouseEvent *);
@@ -24,7 +24,7 @@ public:
 private:
     Ui::CornerDialog *ui;
     Mat img;
-    vector<Point2d> corners;
+    vector<Point2f> corners;
     double rate;
     void calMat(Mat &showmat, Point c, double prerate);
     int showdx;
